Don't log the same IP address twice in a row.
A common pattern is that one host absolutely floods the exim logs
with some sort of spam or relay attempt, and a few other hosts are
active at much slower speeds.

There is little sense in logging the high rate host every single time,
and the latest events page looks a lot nicer when we don't log it
continuously.